Javascript - Refactoring about Utils.js
