+++ assets/src/legacy/map.js
@@ -12,7 +12,7 @@ import { extend } from 'ol/extent.js';
 
 import WFS from '../modules/WFS.js';
 import WMS from '../modules/WMS.js';
-import Utils from '../modules/Utils.js';
+import { Utils } from '../modules/Utils.js';
 
 window.lizMap = function() {
     /**
