+++ assets/src/modules/Popup.js
@@ -8,7 +8,7 @@
 
 import Overlay from 'ol/Overlay.js';
 import WMS from '../modules/WMS.js';
-import Utils from '../modules/Utils.js';
+import { Utils } from './Utils.js';
 
 /**
  * @class
